feat: onboarding was separated
