--- vue/src/views/OnboardingView.vue
@@ -1,7 +1,0 @@
-<template>
-  <onboarding-form></onboarding-form>
-</template>
-
-<script setup lang="ts">
-import OnboardingForm from "@/components/onboarding/OnboardingForm.vue";
-</script>
