+++ vue/src/views/onboarding/OnboardingCompany.vue
@@ -1,0 +1,7 @@
+<template>
+  <company-form></company-form>
+</template>
+
+<script setup lang="ts">
+import CompanyForm from "@/components/onboarding/CompanyForm.vue";
+</script>
