+++ vue/src/views/onboarding/OnboardingGarden.vue
@@ -1,0 +1,7 @@
+<template>
+  <garden-form></garden-form>
+</template>
+
+<script setup lang="ts">
+import GardenForm from "@/components/onboarding/GardenForm.vue";
+</script>
